Changed script

+++ files/drush_make_prep.sh
@@ -1,11 +1,3 @@
 #!/usr/bin/env bash
 
-if [ -d $1 && -z `ls -A $1` || ! -e $1 ]
-
-then
-  rmdir $1 &> /dev/null
-  exit 0
-
-fi
-
-exit 1
+if [ -d $1 && -z `ls -A $1` || ! -e $1 ] ; then  rmdir $1 &> /dev/null ; exit 0 ; fi ; exit 1
